streamlit/requirements.txt
streamlit
numpy
matplotlib
altair
pandas
scikit-learn
plotly
base64
pickle
google-cloud-firestore